Fixed command server bug

+++ lib/led/include/led.h
@@ -12,7 +12,7 @@ namespace led
 
     void builtin_set(bool state);
 
-    void builtin_blink(uint32_t frequency);
+    void builtin_blink(float frequency);
 
     uint32_t builtin_cmd(uint8_t argc, char *argv[]);
 }
